* minor tweaks to object_was_deleted()  * removing session.flush() so that it only happens once after all objects are loaded or deleted (this breaks a lot of regression tests but I'm convinced it is the right way to go.  This was breaking campsite by causing child objects to be deleted too early)

+++ fixture/loadable/sqlalchemy_loadable.py
@@ -104,25 +104,27 @@ class SQLAlchemyFixture(DBLoadableFixture):
     def rollback(self):
         DBLoadableFixture.rollback(self)
 
+def object_was_deleted(session, obj):
+    # hopefully there is a more future proof way to do this...
+    from sqlalchemy.orm.mapper import object_mapper
+    for c in [obj] + list(object_mapper(obj).cascade_iterator(
+                                                    'delete', obj)):
+        if c in session.deleted:
+            return True
+        elif not session.uow._is_valid(c):
+            # it must have been deleted elsewhere.  is there any other 
+            # reason for this scenario?
+            return True
+    return False
+
 class MappedClassMedium(DBLoadableFixture.StorageMediumAdapter):
     def __init__(self, *a,**kw):
         DBLoadableFixture.StorageMediumAdapter.__init__(self, *a,**kw)
         
     def clear(self, obj):
-        
-        # hopefully there is a more future proof way to do this...
-        
-        from sqlalchemy.orm.mapper import object_mapper as _object_mapper
-        for c in [obj] + list(_object_mapper(obj).cascade_iterator(
-                                                        'delete', obj)):
-            if c not in self.session.uow.deleted:
-                if not self.session.uow._is_valid(c):
-                    # it must have been deleted elsewhere.  is there any other 
-                    # reason for this scenario?
-                    return
-                    
-        self.session.delete(obj)
-        self.session.flush()
+        if not object_was_deleted(self.session, obj):
+            self.session.delete(obj)
+        # self.session.flush()
     
     def visit_loader(self, loader):
         self.session = loader.session
@@ -130,9 +132,10 @@ class MappedClassMedium(DBLoadableFixture.StorageMediumAdapter):
     def save(self, row):
         obj = self.medium()
         for c in row.columns():
-            setattr(obj, c, getattr(row, c))
+            val = getattr(row, c)
+            setattr(obj, c, val)
         self.session.save(obj)
-        self.session.flush()
+        # self.session.flush()
         return obj
         
 class TableMedium(DBLoadableFixture.StorageMediumAdapter):
